INT-2729 Redis Outbound Adapter exposed setters for RedisSerializers, modified namespace, added tests
INT-2729 added missing test file
